Crate pravega_client

A Rust client for Pravega.

Pravega is an open source storage system implementing Streams as first-class primitive for storing/serving continuous and unbounded data. It has a number of exciting features including Exactly Once Semantics, Auto Scaling, Unlimited Retention and etc. More details at the website.

Pravega client in Rust provides a few APIs at high level:

Event provides a way to write and read discrete item.

Byte provides a way to write and read raw bytes.

Modules

byte

The Byte API for writing and reading data to a segment in raw bytes.

client_factory

Factory to create components in Pravega Rust client.

error

event

The Event API for writing and reading data.

index

The Index API provides a way to build a monotonic index with a stream.

sync

Pravega synchronization primitives.