Struct pravega_client::byte::reader::ByteReader

pub struct ByteReader { /* fields omitted */ }

A ByteReader enables reading raw bytes from a segment.

The ByteReader implements Read and Seek trait in the standard library.

Internally ByteReader uses a prefetching reader that prefetches data from the server in the background. The prefetched data is cached in memory so any sequential reads should be able to hit the cache.

Any seek operation will invalidate the cache and causes cache miss, so frequent seek and read operations might not have good performance.

You can also wrap ByteReader with BufReader, but doing so will not increase performance further.

Examples

use pravega_client_config::ClientConfigBuilder;
use pravega_client::client_factory::ClientFactory;
use pravega_client_shared::ScopedSegment;
use std::io::Read;

fn main() {
    // assuming Pravega controller is running at endpoint `localhost:9090`
    let config = ClientConfigBuilder::default()
        .controller_uri("localhost:9090")
        .build()
        .expect("creating config");

    let client_factory = ClientFactory::new(config);

    // assuming scope:myscope, stream:mystream and segment with id 0 do exist.
    let segment = ScopedSegment::from("myscope/mystream/0");

    let mut byte_reader = client_factory.create_byte_reader(segment);
    let mut buf: Vec<u8> = vec![0; 4];
    let size = byte_reader.read(&mut buf).expect("read from byte stream");
}

Implementations

impl ByteReader

pub async fn read_async(&mut self, buf: &mut [u8]) -> Result<usize, Error>

Read data asynchronously.

let mut byte_reader = client_factory.create_byte_reader_async(segment).await;
let mut buf: Vec<u8> = vec![0; 4];
let size = byte_reader.read_async(&mut buf).expect("read");

pub async fn current_head(&self) -> Result<u64>

Return the head of current readable data in the segment.

The ByteReader is initialized to read from the segment at offset 0. However, it might encounter the SegmentIsTruncated error due to the segment has been truncated. In this case, application should call this method to get the current readable head and read from it.

let mut byte_reader = client_factory.create_byte_reader_async(segment).await;
let offset = byte_reader.current_head().await.expect("get current head offset");

pub async fn current_tail(&self) -> Result<u64>

Return the tail offset of the segment.

let mut byte_reader = client_factory.create_byte_reader_async(segment).await;
let offset = byte_reader.current_tail().await.expect("get current tail offset");

pub fn current_offset(&self) -> u64

Return the current read offset.

let mut byte_reader = client_factory.create_byte_reader(segment);
let offset = byte_reader.current_offset();

pub fn available(&self) -> usize

Return the bytes that are available to read instantly without fetching from server.

ByteReader has a buffer internally. This method returns the size of remaining data in that buffer.

let mut byte_reader = client_factory.create_byte_reader(segment);
let size = byte_reader.available();

pub async fn seek_async(&mut self, pos: SeekFrom) -> Result<u64>

Seek to an offset asynchronously.