Struct pravega_client::client_factory::ClientFactory

pub struct ClientFactory(_);

Applications should use ClientFactory to create resources they need.

ClientFactory contains a connection pool that is shared by all the readers and writers it creates. It also contains a tokio runtime that is used to drive for async tasks. Spawned tasks in readers and writers are tied to this runtime.

Applications can call clone on ClientFactory in case ownership is needed. It holds an Arc to the internal object so cloned objects are still sharing the same connection pool and runtime.

Implementations

impl ClientFactory

pub fn new(config: ClientConfig) -> ClientFactory

Create a new ClientFactory.

Examples

use pravega_client_config::ClientConfigBuilder;
use pravega_client::client_factory::ClientFactory;

fn main() {
   let config = ClientConfigBuilder::default()
        .controller_uri("localhost:8000")
        .build()
        .expect("create config");
    let client_factory = ClientFactory::new(config);
}

pub fn runtime(&self) -> &Runtime

pub fn config(&self) -> &ClientConfig

pub fn controller_client(&self) -> &dyn ControllerClient

pub fn create_event_writer(&self, stream: ScopedStream) -> EventWriter

pub async fn create_reader_group(
&self,
scope: Scope,
reader_group_name: String,
stream: ScopedStream
) -> ReaderGroup

pub async fn create_transactional_event_writer(
&self,
stream: ScopedStream,
writer_id: WriterId
) -> TransactionalEventWriter

pub fn create_byte_writer(&self, segment: ScopedSegment) -> ByteWriter

pub fn create_byte_reader(&self, segment: ScopedSegment) -> ByteReader

pub async fn create_table(&self, scope: Scope, name: String) -> Table

pub async fn create_synchronizer(
&self,
scope: Scope,
name: String
) -> Synchronizer