Struct pravega_client::event::reader_group_state::ReaderGroupState

pub struct ReaderGroupState { /* fields omitted */ }

ReaderGroupState encapsulates all readers states.

Implementations

impl ReaderGroupState

pub async fn add_reader(
&mut self,
reader: &Reader
) -> Result<(), ReaderGroupStateError>

Adds a reader to the reader group state.

pub async fn check_online(&mut self, reader: &Reader) -> bool

pub async fn get_online_readers(&mut self) -> Vec<Reader>

Returns the active readers in a vector.

pub async fn compute_segments_to_acquire_or_release(
&mut self,
reader: &Reader
) -> isize

Compute the number of segments to acquire.

pub async fn get_segments(&mut self) -> HashSet<ScopedSegment>

Return the list of all segments.

pub async fn assign_segment_to_reader(
&mut self,
reader: &Reader
) -> Result<Option<ScopedSegment>, ReaderGroupStateError>

Assign an unassigned segment to a given reader

pub async fn get_segments_for_reader(
&mut self,
reader: &Reader
) -> Result<HashSet<(ScopedSegment, Offset)>, SynchronizerError>

Return the list of segments assigned to the requested reader.

pub async fn release_segment(
&mut self,
reader: &Reader,
segment: &ScopedSegment,
offset: &Offset
) -> Result<(), ReaderGroupStateError>

Release a currently assigned segment from the given reader.

pub async fn segment_completed(
&mut self,
reader: &Reader,
segment_completed: &ScopedSegment,
successors_mapped_to_their_predecessors: &HashMap<SegmentWithRange, Vec<Segment>>
) -> Result<(), ReaderGroupStateError>

Remove the completed segments and add its successors for next to read. This should be called by the reader who’s reading the current segment. Since a segment cannot be read by multiple readers, we can assume this won’t be called by multiple processors at the same time.