Struct pravega_client::event::transactional_writer::TransactionalEventWriter

pub struct TransactionalEventWriter { /* private fields */ }

Write events to a stream transactionally.

Events that are written to the transaction can be committed atomically, which means that reader cannot see any writes prior to committing and will not see any writes if the transaction is aborted.

Example

use tokio;
use pravega_client_shared::{Timestamp, ScopedStream, Scope, Stream, WriterId, PravegaNodeUri};
use pravega_client::client_factory::ClientFactory;
use pravega_client_config::ClientConfigBuilder;

#[tokio::main]
async fn main() {
    let scope_name = Scope::from("txnScope".to_owned());
    let stream_name = Stream::from("txnStream".to_owned());
    let scoped_stream = ScopedStream {
         scope: scope_name.clone(),
         stream: stream_name.clone(),
     };
    // omit the step to create scope and stream in Pravega

    let config = ClientConfigBuilder::default()
        .controller_uri(PravegaNodeUri::from("tcp://127.0.0.2:9091".to_string()))
        .build()
        .expect("creating config");
    let client_factory = ClientFactory::new(config);
    let mut writer = client_factory
        .create_transactional_event_writer(scoped_stream.clone(), WriterId(0))
        .await;

    // start a transaction
    let mut transaction = writer.begin().await.expect("begin transaction");

    // do something with it
    transaction.write_event(None, String::from("hello").into_bytes()).await.unwrap();

    // commit the transaction
    transaction.commit(Timestamp(0u64)).await;
}

Implementations

impl TransactionalEventWriter

pub async fn begin(
&mut self
) -> Result<Transaction, TransactionalEventWriterError>

This method opens a transaction by sending a request to Pravega controller.

pub async fn get_txn(
&self,
txn_id: TxId
) -> Result<Transaction, TransactionalEventWriterError>

This method returns the Transaction based on the given transaction id. If the current transaction is not in open status, meaning it has been committed or aborted, this method will create a closed transaction that only contains the meta data of this transaction.