Struct pravega_client::event_reader::EventReader

pub struct EventReader { /* fields omitted */ }

This represents an event reader. An event reader fetches data from its assigned segments as a SegmentSlice, where a SegmentSlice represents data from a Pravega Segment. It provides the following APIs.

A method to initialize the event reader EventReader#init

A method to obtain a SegmentSlice to read events from a Pravega segment. The user can use the SegmentSlice's iterator API to fetch individual events from a given Segment Slice. EventReader#acquire_segment.

A method to release the Segment back at the given offset. EventReader#release_segment_at. This method needs to be invoked only the user does not consume all the events in a SegmentSlice.

A method to mark the reader as offline.EventReader#reader_offline. This method ensures the segments owned by this readers are transferred to other readers in the reader group.

An example usage pattern is as follows

use pravega_client_config::{ClientConfigBuilder, MOCK_CONTROLLER_URI};
use pravega_client::client_factory::ClientFactory;
use pravega_client_shared::{ScopedStream, Scope, Stream};

#[tokio::main]
async fn main() {
   let config = ClientConfigBuilder::default()
        .controller_uri(MOCK_CONTROLLER_URI)
        .build()
        .expect("creating config");
    let client_factory = ClientFactory::new(config);
    let stream = ScopedStream {
        scope: Scope::from("scope".to_string()),
        stream: Stream::from("stream".to_string()),
    };
    // Create a reader group to read data from the Pravega stream.
    let rg = client_factory.create_reader_group("rg".to_string(), stream).await;
    // Create a reader under the reader group. The segments of the stream are assigned among the
    // readers which are part of the reader group.
    let mut reader1 = rg.create_reader("r1".to_string()).await;
    // read all events from a given segment slice.
    if let Some(mut segment_slice) =  reader1.acquire_segment().await {
        while let Some(event) = segment_slice.next() {
            println!("Event read is {:?}", event);
        }
    }
    // read one event from the a given  segment slice and return it back.
    if let Some(mut segment_slice) = reader1.acquire_segment().await {
        if let Some(event) = segment_slice.next() {
            println!("Event read is {:?}", event);
            // release the segment slice back to the reader.
            reader1.release_segment(segment_slice);
        }
    }
}

Implementations

impl EventReader

pub fn new(
id: Reader,
factory: ClientFactory,
rx: Receiver<SegmentReadResult>,
tx: Sender<SegmentReadResult>,
meta: ReaderState,
rg_state: Arc<Mutex<ReaderGroupState>>
) -> Self

Constructs a new EventReader.

impl EventReader

pub async fn init_reader(
id: String,
rg_state: Arc<Mutex<ReaderGroupState>>,
factory: ClientFactory
) -> Self

Initialize the reader. This fetches the assigned segments from the TableSynchronizer and spawns background tasks to start reads from those Segments.

pub fn release_segment(&mut self, slice: SegmentSlice)

Release a partially read segment slice back to event reader.

pub fn release_segment_at(&mut self, slice: SegmentSlice, offset: i64)

Release a segment back to the reader and also indicate the offset upto which the segment slice is consumed.

pub async fn reader_offline<'_>(&'_ mut self)

Mark the reader as offline. This will ensure the segments owned by this reader is distributed to other readers in the ReaderGroup.

pub async fn acquire_segment<'_>(&'_ mut self) -> Option<SegmentSlice>

This function returns a SegmentSlice from the data received from the SegmentStore(s). Individual events can be read from the data received using SegmentSlice.next().

Invoking this function multiple times ensure multiple SegmentSlices corresponding to different Segments of the stream are received. In-case we receive data for an already acquired SegmentSlice this method waits until SegmentSlice is completely consumed before returning the data.