Trait pravega_client::segment_reader::AsyncSegmentReader

pub trait AsyncSegmentReader: Send + Sync {
#[must_use]    fn read<'life0, 'async_trait>(
        &'life0 self, 
        offset: i64, 
        length: i32
    ) -> Pin<Box<dyn Future<Output = StdResult<SegmentReadCommand, ReaderError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

AsyncSegmentReader is used to read from a given segment given the connection pool and the Controller URI The reads given the offset and the length are processed asynchronously. e.g: usage pattern is AsyncSegmentReaderImpl::new(&segment_name, connection_pool, "http://controller uri").await

Required methods

#[must_use]fn read<'life0, 'async_trait>(
&'life0 self,
offset: i64,
length: i32
) -> Pin<Box<dyn Future<Output = StdResult<SegmentReadCommand, ReaderError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...