Struct pravega_client::tablemap::TableMap

pub struct TableMap { /* fields omitted */ }

Implementations

impl TableMap

pub async fn new(
scope: Scope,
name: String,
factory: ClientFactory
) -> Result<TableMap, TableError>

create a table map

pub async fn get<K, V>(&self, k: &K) -> Result<Option<(V, Version)>, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Returns the latest value corresponding to the key.

If the map does not have the key None is returned. The version number of the Value is returned by the API.

pub async fn insert<K, V>(
&self,
k: &K,
v: &V,
offset: i64
) -> Result<Version, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Unconditionally inserts a new or update an existing entry for the given key. Once the update is performed the newer version is returned.

pub async fn insert_conditionally<K, V>(
&self,
k: &K,
v: &V,
key_version: Version,
offset: i64
) -> Result<Version, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Conditionally inserts a key-value pair into the table map. The Key and Value are serialized to bytes using cbor.

The insert is performed after checking the key_version passed. Once the update is done the newer version is returned. TableError::BadKeyVersion is returned in case of an incorrect key version.

pub async fn remove<K: Serialize + DeserializeOwned>(
&self,
k: &K,
offset: i64
) -> Result<(), TableError>

Unconditionally remove a key from the Tablemap. If the key does not exist an Ok(()) is returned.

pub async fn remove_conditionally<K>(
&self,
k: &K,
key_version: Version,
offset: i64
) -> Result<(), TableError> where
K: Serialize + DeserializeOwned,

Conditionally remove a key from the Tablemap if it matches the provided key version. TableError::BadKeyVersion is returned incase the version does not exist.

pub async fn get_all<K, V>(
&self,
keys: Vec<&K>
) -> Result<Vec<Option<(V, Version)>>, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Returns the latest values for a given list of keys. If the tablemap does not have a key a None is returned for the corresponding key. The version number of the Value is also returned by the API

pub async fn insert_all<K, V>(
&self,
kvps: Vec<(&K, &V)>,
offset: i64
) -> Result<Vec<Version>, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Unconditionally inserts a new or updates an existing entry for the given keys. Once the update is performed the newer versions are returned.

pub async fn insert_conditionally_all<K, V>(
&self,
kvps: Vec<(&K, &V, Version)>,
offset: i64
) -> Result<Vec<Version>, TableError> where
K: Serialize + DeserializeOwned,
V: Serialize + DeserializeOwned,

Conditionally inserts key-value pairs into the table map. The Key and Value are serialized to to bytes using cbor

The insert is performed after checking the key_version passed, in case of a failure none of the key-value pairs are persisted. Once the update is done the newer version is returned. TableError::BadKeyVersion is returned in case of an incorrect key version.

pub async fn remove_all<K>(
&self,
keys: Vec<&K>,
offset: i64
) -> Result<(), TableError> where
K: Serialize + DeserializeOwned,

Unconditionally remove the provided keys from the table map.

pub async fn remove_conditionally_all<K>(
&self,
keys: Vec<(&K, Version)>,
offset: i64
) -> Result<(), TableError> where
K: Serialize + DeserializeOwned,

Conditionally remove keys after checking the key version. In-case of a failure none of the keys are removed.

pub fn read_keys_stream<'stream, 'map: 'stream, K: 'stream>(
&'map self,
max_keys_at_once: i32
) -> impl Stream<Item = Result<(K, Version), TableError>> + 'stream where
K: Serialize + DeserializeOwned + Unpin,

Read keys as an Async Stream. This method deserializes the Key based on the type.

pub fn read_entries_stream<'stream, 'map: 'stream, K: 'map, V: 'map>(
&'map self,
max_entries_at_once: i32
) -> impl Stream<Item = Result<(K, V, Version), TableError>> + 'stream where
K: Serialize + DeserializeOwned + Unpin,
V: Serialize + DeserializeOwned + Unpin,

Read entries as an Async Stream. This method deserialized the Key and Value based on the inferred type.

pub fn read_entries_stream_from_position<'stream, 'map: 'stream, K: 'map, V: 'map>(
&'map self,
max_entries_at_once: i32,
from_position: i64
) -> impl Stream<Item = Result<(K, V, Version, i64), TableError>> + 'stream where
K: Serialize + DeserializeOwned + Unpin,
V: Serialize + DeserializeOwned + Unpin,

Read entries as an Async Stream from a given position. This method deserialized the Key and Value based on the inferred type.