Crate pravega_client

A Rust client for Pravega.

Pravega is an open source storage system implementing Streams as first-class primitive for storing/serving continuous and unbounded data. It has a number of exciting features including Exactly Once Semantics, Auto Scaling, Unlimited Retention and etc. More details at the website.

Pravega client in Rust provides a few APIs at high level:

EventStreamWriter and EventStreamReader provide a way to write and read discrete item.

ByteStream API provides a way to write and read raw bytes.

Transaction API provides a mechanism for writing many events atomically.

Modules

byte_stream

client_factory

error

event_reader

event_reader_group

event_stream_writer

metric

raw_client

reader_group

reader_group_config

segment_metadata

segment_reader

segment_slice

table_synchronizer

tablemap

trace

transaction

Macros

update