Module pravega_client::byte

The Byte API for writing and reading data to a segment in raw bytes.

Byte streams are restricted to single-segment streams, but, unlike event streams, there is no event framing, so a reader cannot distinguish separate writes for you, and, if needed, you must do so by convention or by protocol in a layer above the reader.

Note

Sharing a single-segment stream between the byte API and the Event readers/writers will CORRUPT YOUR DATA in an unrecoverable way.

Sharing a single-segment stream between multiple byte writers is possible but it might generate interleaved data.

Modules

reader

writer

Structs

ByteReader

A ByteReader enables reading raw bytes from a segment.

ByteWriter

Allow for writing raw bytes directly to a segment.